Removes ParameterizedQuery usage
This commit cleans up some code. Fixes #96. Uses arrow functions for
callbacks. The commit refactors the usage for the pg-promise API to use
`any` instead of `many` and uses the regular query instead of
`ParameterizedQuery`.

The commit also cleans up some usage and replaces uses of `every` with a
for-loop or `find` where appropriate.